+++ NEWS.md
@@ -11,6 +11,7 @@
 ## Minor changes
 * Added `conda.path` in `ExportSeurat` to specify conda enviroment used.
 * Fixed bug in `SCELoom` to keep gene names and cell ID.
+* `Bam2Fastq` supported normal bam files (non-10x bam files).
 
 -------------------